% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmrf.R
\name{createSpatialNetwork}
\alias{createSpatialNetwork}
\title{createSpatialNetwork}
\usage{
createSpatialNetwork(coords, k = 4)
}
\arguments{
\item{coords}{Input a named list where X element contains
x coordinates and Y element contains y coordinates.}

\item{k}{Specify number of initial clusters to generate.
Default \code{4}.}
}
\value{
A data.table containg the spatial network.
}
\description{
Creates a KNN based spatial network from
input spatial coordinates.
}
