% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmrf.R
\name{initializeHMRF}
\alias{initializeHMRF}
\title{initializeHMRF}
\usage{
initializeHMRF(
  y,
  spatial_network,
  spatial_genes,
  k = 3,
  hmrf_seed = 100,
  nstart = 1000,
  factor_step = 1.05,
  tolerance = 1e-05
)
}
\arguments{
\item{y}{Input expression matrix where rows are cells
and columns are features.}

\item{spatial_network}{Input a spatial network previously
generated from \code{\link{createSpatialNetwork}} function.}

\item{spatial_genes}{A character vector of names of genes
to use with HMRF. It is recommended to use only the top
most spatially variable genes at this step.}

\item{k}{Specify a k value which defines the number of
spatial domains to identify.}

\item{hmrf_seed}{Set a seed value}

\item{nstart}{Set a nstart value}

\item{factor_step}{Set a factor_step value}

\item{tolerance}{Set a tolerance value}
}
\value{
A list of initialized parameters for HMRF.
}
\description{
Initializes the parameters of the HMRF using
the expression matrix and the spatial network previously
computed.
}
