% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmrf.R
\name{plotHMRF}
\alias{plotHMRF}
\title{plotHMRF}
\usage{
plotHMRF(spatial_network, HMRFoutput)
}
\arguments{
\item{spatial_network}{Input the spatial network.}

\item{HMRFoutput}{Input the HMRF results.}
}
\value{
A list of plots.
}
\description{
Plots the spatial domains
identified previously on a 2D spatial plot.
}
