% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reader.R
\name{readCoordinates}
\alias{readCoordinates}
\title{Reads spatial coordinates from a text file}
\usage{
readCoordinates(filePath, header = FALSE)
}
\arguments{
\item{filePath}{Specify the file path for the input
coordinates file.}

\item{header}{Specify if the input file has a header.
Default \code{FALSE}.}
}
\value{
A list of 2 elements where the first one
is data.frame of X & Y coordinates and the second
element of the list is a numeric vector of FOV.
}
\description{
This functions reads in spatial coordinates
and fov from a text file. It assumes that text file has no
header and that there are three columns and as many
rows as there are number of samples/spots/cells. The
input text file has four columns. First column are just
the indices for the rows which will be ignored while
reading in the text file. Second column is the
FOV (field of vision) value for a particular sample.
The third and fourth columns are the X & Y coordinates for
the respective sample. The output is a list of two
elements. First element accessed by $coords are
the coordinates and the second element is the fov
accessed by $field.
}
