% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reader.R
\name{readGenes}
\alias{readGenes}
\title{Reads names of genes/features from a text file}
\usage{
readGenes(filePath, header = FALSE)
}
\arguments{
\item{filePath}{Specify the file path of the file
containing the gene or feature names.}

\item{header}{Specify if the input file has a header.
Default \code{FALSE}.}
}
\value{
A character vector for gene or feature names.
}
\description{
This function reads in genes or feature names
from a text file and outputs a vector of these names.
The function assumes that each gene name is one
line in the text file.
}
