% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmrf.R
\name{runHMRF}
\alias{runHMRF}
\title{runHMRF}
\usage{
runHMRF(
  y,
  nei,
  numnei,
  blocks,
  damp,
  mu,
  sigma,
  k,
  genes,
  edgelist,
  beta_initial_value = 0,
  beta_increment_value = 10,
  beta_iterations = 5
)
}
\arguments{
\item{y}{Input expression matrix where rows are cells
and columns are features.}

\item{nei}{Specify the neighbours against each cell.}

\item{numnei}{Specify the number of neighbours
against each cell.}

\item{blocks}{Specify the groups of clusters computed
by a graph coloring algorithm where each no two
adjcacent cells have same colors.}

\item{damp}{Set damp param}

\item{mu}{Set mu param}

\item{sigma}{Set sigma param}

\item{k}{Set k value that defines the number of
spatial domains to identify.}

\item{genes}{Specify list of genes}

\item{edgelist}{Specify all edges between cells.}

\item{beta_initial_value}{Specify initial beta value.}

\item{beta_increment_value}{Specify increment value.}

\item{beta_iterations}{Specify total iterations.}
}
\value{
A list containing results of HMRF.
}
\description{
runHMRF
}
